* origin/feature/transactions:
  PVM File changes
  Committing ISO500Settlement Code changes
  Changes added after merger - Initialization.
  CbaseNode
  PVM Code Changes
  Corrected GlobalData APIs.
  Changes committed related to ISO320Initialization.
  Receive Complete Packet added in TCP
  activation completed
  Added changes related to ChargeSlip and FileSystem Fixes.
  CRC32 File deleted no need
  Activation running code added ISOMessage File modified

+++ ePOS/Modules/Transactions/Classes/LOGIN_ACCOUNTS.swift
@@ -13,10 +13,10 @@ class LOGINACCOUNTS: Codable
     var m_strUserID: String = ""
     var m_strPIN: String = ""
     var m_strCreatedBy: String = ""
-    var m_sAccountType: Int8 = 0
+    var m_sAccountType: Int16 = 0
     var m_strCreatedOn: String = ""
     var m_strUpdatedOn: String = ""     //refactor to UpdatedOn
     var m_strLastLoginOn: String = ""
     var m_strUUID: String = ""
-    var m_sFlagChangeType: Int8 = 0 //0 Nothing 1 New 2 Modify 3 Delete
+    var m_sFlagChangeType: Int16 = 0 //0 Nothing 1 New 2 Modify 3 Delete
 }
